BioVerify/BioVerify/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "google-api-python-client>=2.179.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "gradio>=5.42.0",
    "jinja2>=3.1.6",
    "joblib>=1.5.1",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "plotly>=6.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "qrcode[pil]>=8.2",
    "requests>=2.32.4",
    "scikit-learn>=1.7.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
